route entity -- move entity to org.mobilitydata.gtfsvalidator.domain.entity.gtfs package
