+++ domain/src/main/java/org/mobilitydata/gtfsvalidator/domain/entity/gtfs/RouteType.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.mobilitydata.gtfsvalidator.domain.entity;
+package org.mobilitydata.gtfsvalidator.domain.entity.gtfs;
 
 import java.util.stream.Stream;
